Refactor code and structure
