+++ src/App.js
@@ -1,7 +1,7 @@
 import "./styles.css";
-import Nav from "./Nav";
-import { useScreen } from "./changeScreen";
-import { UrlProvider } from "./useUrl";
+import Nav from "./components/Nav";
+import { useScreen } from "./context/changeScreen";
+import { UrlProvider } from "./context/useUrl";
 export default function App() {
   const { screen } = useScreen();
   return (
