+++ src/index.js
@@ -2,8 +2,8 @@ import { StrictMode } from "react";
 import ReactDOM from "react-dom";
 
 import App from "./App";
-import { CartProvider } from "./cartContext";
-import { ScreenProvider } from "./changeScreen";
+import { CartProvider } from "./context/cartContext";
+import { ScreenProvider } from "./context/changeScreen";
 const rootElement = document.getElementById("root");
 ReactDOM.render(
   <StrictMode>
